CGI: integrating cgi to core logic : cgi works with both GET & POST . need to be tested more

+++ src/server/server_cgi.cpp
@@ -163,7 +163,7 @@ void    Server::processCgiWriteEvent(int cgi_pipe) // DANGER : reference may be 
                     }
         
         close(cgi_client.cgi->script_in[1]);
-        cgi_client.cgi->script_in[1] = -1;
+        // cgi_client.cgi->script_in[1] = -1;
 
         cgi_client.cgi->cgi_status = CGI_READING;
         std::cout << "cgi done writing : " << cgi_pipe << std::endl;
@@ -180,7 +180,11 @@ void    Server::processCgiWriteEvent(int cgi_pipe) // DANGER : reference may be 
 
 void    Server::close_cgi_client(int fd)
 {
+    if (fd == -1) // of leaks this is the cause . check if script_in[1] was seted to -1 after cgi_client created
+        return ;
+    std::cout << "close cgi client " << fd << std::endl;
     CgiClient& cgi_client = cgi_clients.at(fd);
+    std::cout << "close cgi client " << fd << std::endl;
 
     kill(cgi_client.cgi->pid, SIGKILL);
     waitpid(cgi_client.cgi->pid, NULL, WNOHANG);
